+++ R/scratch.R
@@ -156,7 +156,5 @@
 
 # writeLines(mapdeck::geojson, "~/Desktop/geo_melbourne.json")
 
-#s
-# sf <- geojsonsf:::rcpp_read_sf_file_stream( file = normalizePath("~/Desktop/geo_melbourne.json"), FALSE )
 # sf <- geojsonsf::geojson_sf( "~/Desktop/geo_melbourne.json" )
 # sfc <- geojsonsf::geojson_sfc( "~/Desktop/geo_melbourne.json" )