+++ src/read_geojson.cpp
@@ -1,7 +1,6 @@
 #include <Rcpp.h>
 #include <rapidjson/document.h>
 
-#include <rapidjson/istreamwrapper.h>
 #include <rapidjson/filereadstream.h>
 
 #include "geojsonsf/geojsonsf.h"
@@ -11,46 +10,47 @@
 #include "geojsonsf/geojson/geojson_properties.hpp"
 #include <fstream>
 
-Rcpp::StringVector buffer_string(std::string file) {
-  std::ifstream ifs(file);
-  std::stringstream buffer;
-  buffer << ifs.rdbuf();
-  return buffer.str();
-}
+// Rcpp::StringVector buffer_string(std::string file) {
+//   std::ifstream ifs(file);
+//   std::stringstream buffer;
+//   buffer << ifs.rdbuf();
+//   return buffer.str();
+// }
 
 
-rapidjson::Document buffer_string( const char* file ) {
+rapidjson::Document buffer_string( const char* file, const char* mode, int buffer_size = 1024 ) {
 
-	FILE* fp = fopen(file, "rb"); // TODO - windows uses "r"
-	char readBuffer[1024];
-	size_t buffer_size = sizeof( readBuffer );
-	//Rcpp::Rcout << "buffer_size: " << buffer_size << std::endl;
-	rapidjson::FileReadStream is(fp, readBuffer, buffer_size );
+	FILE* fp = fopen(file, mode );
+	char readBuffer[ buffer_size ];
 
-	//std::ifstream ifs( file );
-	//rapidjson::IStreamWrapper isw( ifs );
+	rapidjson::FileReadStream is(fp, readBuffer, sizeof( readBuffer ) );
 
 	Document d;
 	d.ParseStream( is );
-
-	size_t s = d.Size();
-	//Rcpp::Rcout << "(rcpp_read_sf_file_stream) doc size: " << s << std::endl;
-
 	fclose(fp);
 
 	return d;
 }
 
 // [[Rcpp::export]]
-Rcpp::List rcpp_read_sf_file( const char* file, bool flatten_geometries ) {
-
-	rapidjson::Document d = buffer_string( file );
+Rcpp::List rcpp_read_sf_file(
+		const char* file,
+		const char* mode,
+		bool flatten_geometries,
+		int buffer_size = 1024
+	) {
+	rapidjson::Document d = buffer_string( file, mode, buffer_size );
   return geojsonsf::sf::generic_geojson_to_sf( d , flatten_geometries );
 }
 
 // [[Rcpp::export]]
-Rcpp::List rcpp_read_sfc_file( const char* file, bool flatten_geometries) {
-	rapidjson::Document d = buffer_string( file );
+Rcpp::List rcpp_read_sfc_file(
+		const char* file,
+		const char* mode,
+		bool flatten_geometries,
+		int buffer_size = 1024
+	) {
+	rapidjson::Document d = buffer_string( file, mode, buffer_size );
 	return geojsonsf::sf::create_sfc( d, flatten_geometries);
 }